Clean up the auth-manager and add proper README.
