+++ no-waste-auth-manager/src/main/kotlin/edu/ubb/micro/nowaste/authmanager/dto/UserDTO.kt
@@ -2,6 +2,9 @@ package edu.ubb.micro.nowaste.authmanager.dto
 
 import com.fasterxml.jackson.annotation.JsonProperty
 
+/**
+ * Represents a User with basic fields.
+ */
 data class UserDTO(
 	@JsonProperty("id") val id: String,
 	@JsonProperty("username") val userName: String,
