+++ no-waste-auth-manager/src/main/kotlin/edu/ubb/micro/nowaste/authmanager/service/ServiceException.kt
@@ -1,9 +1,10 @@
 package edu.ubb.micro.nowaste.authmanager.service
 
 /**
- * [RuntimeException] which can be thrown by Service layer.
+ * [RuntimeException] representing an error on Service Layer.
  */
 class ServiceException : RuntimeException {
+	constructor() : super()
 	constructor(message: String) : super(message)
 	constructor(message: String, throwable: Throwable) : super(message, throwable)
