+++ no-waste-auth-manager/src/main/kotlin/edu/ubb/micro/nowaste/authmanager/service/UserAuthUseCaseImpl.kt
@@ -1,6 +1,7 @@
 package edu.ubb.micro.nowaste.authmanager.service
 
 import edu.ubb.micro.nowaste.authmanager.model.User
+import edu.ubb.micro.nowaste.authmanager.service.auth.AuthService
 import edu.ubb.micro.nowaste.authmanager.service.user.UserService
 import org.springframework.stereotype.Component
 import javax.servlet.http.HttpServletRequest
