+++ no-waste-auth-manager/src/main/kotlin/edu/ubb/micro/nowaste/authmanager/service/user/UserService.kt
@@ -5,6 +5,11 @@ import edu.ubb.micro.nowaste.authmanager.service.ServiceException
 
 interface UserService {
 
+	/**
+	 * Retrieves from source the [User] identified by [userId] id.
+	 *
+	 * @return User who's identifier is [userId].
+	 */
 	@Throws(ServiceException::class)
 	fun getUserById(userId: String): User
 }
