+++ no-waste-auth-manager/src/main/kotlin/edu/ubb/micro/nowaste/authmanager/service/user/UserServiceImpl.kt
@@ -2,7 +2,7 @@ package edu.ubb.micro.nowaste.authmanager.service.user
 
 import edu.ubb.micro.nowaste.authmanager.model.User
 import edu.ubb.micro.nowaste.authmanager.service.ServiceException
-import edu.ubb.micro.nowaste.authmanager.service.UserRepository
+import edu.ubb.micro.nowaste.authmanager.repository.UserRepository
 import org.springframework.dao.DataAccessException
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.data.rest.webmvc.ResourceNotFoundException
